Вынес базовый класс в отдельный файл и добавил явное наследование
